requirements.txt
# EDM Shuffle CrewAI Dependencies
# Core CrewAI framework and tools
crewai>=0.28.0
crewai-tools>=0.1.6

# Environment and configuration
python-dotenv>=1.0.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing and utilities  
pandas>=2.0.0
numpy>=1.24.0

# Database integration
supabase>=1.0.0

# Audio processing (TODO: Select based on final implementation)
# librosa>=0.10.0
# pydub>=0.25.0
# soundfile>=0.12.0

# 3D and graphics (TODO: Select based on final implementation)  
# pillow>=10.0.0
# opencv-python>=4.8.0

# Analytics and monitoring
# psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0